+++ tools/autoAnalyseGame/tool.js
@@ -6,7 +6,7 @@
         name: 'autoAnalyseGame',
         category: 'play',
         type: 'multiple',
-        possibleValues: ['loss', 'draw', 'win'],
+        possibleValues: ['loss', 'draw', 'win', 'showRequestAnalysis'],
         defaultValue: false,
         advanced: true
       }
@@ -18,14 +18,18 @@
         'options.autoAnalyseGame': 'Go to Analysis on game end',
         'autoAnalyseGame.loss': 'Loss',
         'autoAnalyseGame.draw': 'Draw',
-        'autoAnalyseGame.win': 'Win'
+        'autoAnalyseGame.win': 'Win',
+        'autoAnalyseGame.showRequestAnalysis': 'Request analysis button',
+        'requestAComputerAnalysis': 'Request a computer analysis'
       },
       'ro-RO': {
         'options.play': 'Joc',
         'options.autoAnalyseGame': 'Intr\u0103 \u00een Analiz\u0103 c\u00e2nd se termin\u0103 jocul',
         'autoAnalyseGame.loss': 'Pierdere',
         'autoAnalyseGame.draw': 'Egal',
-        'autoAnalyseGame.win': 'C\u00e2\u015ftig'
+        'autoAnalyseGame.win': 'C\u00e2\u015ftig',
+        'autoAnalyseGame.showRequestAnalysis': 'Buton analiz\u0103 calculator',
+        'requestAComputerAnalysis': 'Cere o analiz\u0103 f\u0103cut\u0103 de calculator'
       }
     }
 
@@ -33,6 +37,7 @@
     checkEndGame = (ev) => {
       const lt = this.lichessTools;
       const $ = lt.$;
+      const trans = lt.translator;
       if (!$('body').is('.playing')) {
         this.retries = 0;
         return;
@@ -44,19 +49,45 @@
         }
         return;
       }
+      const userId = lt.getUserId();
       let outcome = 'draw';
       if (ev.winner) {
         const winnerHref = $('.game__meta__players .player.'+ev.winner+' a.user-link').attr('href');
         const m = /\/@\/(?<userId>[^\/\?#]+)/.exec(winnerHref);
-        outcome = m?.groups?.userId?.toLowerCase() == lt.getUserId()?.toLowerCase()
+        outcome = m?.groups?.userId?.toLowerCase() == userId?.toLowerCase()
           ? 'win'
           : 'loss';
       }
       this.retries = 0;
       if (this.options[outcome]) {
         const href = $('a.fbt.analysis').attr('href');
-        if (!href) return;
-        lt.global.location.href = href;
+        if (href) {
+          lt.global.location.href = href;
+        }
+      } else {
+        const button = $('.lichessTools-requestAnalysis');
+        if (userId && this.options.showRequestAnalysis) {
+          if (!button.length) {
+            $('<button type="button" class="button text lichessTools-requestAnalysis">')
+              .append($('<span class="is3 text">')
+                        .attr('data-icon',lt.icon.BarChart)
+                        .text(trans.noarg('requestAComputerAnalysis'))
+              )
+              .on('click',async (ev)=>{
+                const href = $('a.fbt.analysis').attr('href');
+                if (!href) return;
+                const m = /^\/(?<id>[^\/?#\s]+)/.exec(href);
+                if (m) {
+                  await lt.net.fetch(`/${m.groups.id}/request-analysis`,{ method: 'POST' });
+                }
+                lt.storage.set('analysis.panel','computer-analysis', { raw: true });
+                lt.global.location.href = href;
+              })
+              .appendTo('.round__side');
+          }
+        } else {
+          button.remove();
+        }
       }
     };
 
@@ -75,7 +106,8 @@
         loss: lt.isOptionSet(value, 'loss'),
         draw: lt.isOptionSet(value, 'draw'),
         win: lt.isOptionSet(value, 'win'),
-        get isSet() { return this.loss || this.draw || this.win }
+        showRequestAnalysis: lt.isOptionSet(value, 'showRequestAnalysis'),
+        get isSet() { return this.loss || this.draw || this.win || this.showRequestAnalysis }
       };
       if (!this.options.isSet) return;
       lt.uiApi.socket.events.on('endData', this.checkEndGame);